add network compatability for freezing
