+++ Assets/Scipts/Core/Ships/ShipMovement.cs
@@ -15,6 +15,15 @@ namespace Scipts.Core.Ships
         public float speed = 11f;
         public float rotationSpeed = 0.25f;
 
+        [SyncVar] 
+        private bool _isFreezed;
+        public bool isFreezed
+        {
+            get { return _isFreezed; }
+            set { _isFreezed = value; CmdChangeIsFreezed(value); }
+        }
+        
+  
         public void Start()
         {
             _playerInput = GetComponent<IPlayerInput>();
@@ -24,13 +33,21 @@ namespace Scipts.Core.Ships
 
         public void Update()
         {
-            _physicsLink.AddForce(_playerInput.Acceleration * speed * transform.forward);
+            if (!isFreezed){
+                _physicsLink.AddForce(_playerInput.Acceleration * speed * transform.forward);
 
-            _transform.Rotate(
-                0f,
-                rotationSpeed * _playerInput.Rotation * Mathf.Min(_physicsLink.Velocity.magnitude, 1f),
-                0f
-            );
+                _transform.Rotate(
+                    0f,
+                    rotationSpeed * _playerInput.Rotation * Mathf.Min(_physicsLink.Velocity.magnitude, 1f),
+                    0f
+                );
+            }
         }
+
+        [Command]
+        public void CmdChangeIsFreezed(bool value){
+            _isFreezed = value;
+        }
+
     }
 }
